Split widget: chart placeholder
